employee entityの戻り値を修正

+++ RingishoSystem/src/main/java/com/kadai/domain/entity/Employee.java
@@ -53,7 +53,7 @@ public class Employee implements UserDetails {
 		this.employeePassword = employeePassword;
 	}
 	
-	public long getEmployeeId() {
+	public int getEmployeeId() {
 		return employeeId;
 	}
